% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_player_stats.R
\name{espn_player_stats}
\alias{espn_player_stats}
\title{Estadisticas de los jugadores de la MLB en ESPN}
\usage{
espn_player_stats(
  year = year_actual,
  stats = "batting",
  season_type = "regular"
)
}
\arguments{
\item{year}{numeric}

\item{stats}{character}

\item{season_type}{character}
}
\value{
Estadisticas de la MLB en ESPN
}
\description{
Estadisticas de los jugadores de la MLB en ESPN
}
\examples{
# year = year que queremos visualizar
# stats = batting, pitching, fielding
# season_type = regular o playoffs
# Get las estadisticas de pitcheo en playoffs de 2003
espn_player_stats(2015, "pitching", "regular")



}
