% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_league_stats.R
\name{get_league_stats}
\alias{get_league_stats}
\title{League stats for MLB from ESPN(baseball)}
\usage{
get_league_stats()
}
\value{
stats & logo
}
\description{
MLB Profesional league baseball data table , stats
}
\examples{
\donttest{
get_league_stats()
}

}
