% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mlb_players.R
\name{get_mlb_players}
\alias{get_mlb_players}
\title{Photos & Logos a& colors for MLB (baseball)}
\usage{
get_mlb_players()
}
\value{
Photos, logos and colors
}
\description{
MLB Profesional league baseball data table with Photos, logos and colors
}
\examples{
\donttest{get_mlb_players()}

}
