% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mlb_teams.R
\name{get_mlb_teams}
\alias{get_mlb_teams}
\title{Logos & colors for MLB (baseball)}
\usage{
get_mlb_teams()
}
\value{
Logos and colors
}
\description{
MLB Profesional league baseball data table , logos and colors
}
\examples{
\donttest{get_mlb_teams()}

}
