% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mlb_wordmarks.R
\name{get_mlb_wordmarks}
\alias{get_mlb_wordmarks}
\title{Wordmarks & colors for MLB (baseball)}
\usage{
get_mlb_wordmarks()
}
\value{
Wordmarks and colors
}
\description{
MLB Profesional league baseball data table , Wordmarks and colors
}
\examples{
\donttest{get_mlb_wordmarks()}

}
