% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_png_logos.R
\name{get_png_logos}
\alias{get_png_logos}
\title{4 png Logos & colors for MLB (baseball)}
\usage{
get_png_logos()
}
\value{
Logos and colors
}
\description{
MLB Profesional league baseball data table , logos and colors
}
\examples{


\donttest{get_png_logos()}

}
