% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_players_mlb.R
\name{get_reference_players_mlb}
\alias{get_reference_players_mlb}
\title{Estadisticas de los jugadores de la MLB en Baseball Reference}
\usage{
get_reference_players_mlb(
  year = year_actual,
  stats = "batting",
  type = "standard"
)
}
\arguments{
\item{year}{numeric}

\item{stats}{character}

\item{type}{character}
}
\value{
Estadisticas de la MLB en baseball-reference.com
}
\description{
Estadisticas de los jugadores de la MLB en Baseball Reference
}
\examples{
# year = year que queremos visualizar
# stats = batting, pitching, fielding
# type = 'standard', 'appearances', 'pitcher', 'catcher', 'firstbase', 'secondbase',
# 'thirdbase',  'shortstop', 'leftfield', 'centerfield', 'rightfield', 'outfield'
# 'advanced', 'value', 'probability', 'ratio', 'baserunning', 'batting', 'battingagainst',
# 'startingpitching',
# Get las estadisticas de pitcheo de 1903
# get_reference_playeres_mlb (1903, "pitching", "standard")


}
