% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_team_mlb.R
\name{get_reference_team_mlb}
\alias{get_reference_team_mlb}
\title{Estadisticas de los equipos de la MLB en Baseball Reference}
\usage{
get_reference_team_mlb(
  year = year_actual,
  stats = "batting",
  type = "standard"
)
}
\arguments{
\item{year}{numeric}

\item{stats}{character}

\item{type}{character}
}
\value{
Estadisticas de la MLB en baseball-reference.com
}
\description{
Estadisticas de los equipos de la MLB en Baseball Reference
}
\examples{
# year = year que queremos visualizar
# stats = batting, pitching, fielding
# type = 'standard', 'appearances', 'pitcher', 'catcher', 'firstbase', 'secondbase',
# 'thirdbase',  'shortstop', 'leftfield', 'centerfield', 'rightfield', 'outfield'
# 'standard', 'advanced', 'value', 'probability', 'ratio', 'baserunning', 'batting',
# 'battingagainst', 'startingpitching'
#
# Get las estadisticas de pitcheo de 1873
 \donttest{get_reference_team_mlb(2015, "pitching", "standard")}


}
