% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_team_standings.R
\name{get_reference_team_standings}
\alias{get_reference_team_standings}
\title{Clasificaciones de los equipos de la MLB en Baseball Reference}
\usage{
get_reference_team_standings(year = year_actual)
}
\arguments{
\item{year}{numeric}
}
\value{
Estadisticas de la MLB en baseball-reference.com
}
\description{
Clasificaciones de los equipos de la MLB en Baseball Reference
}
\examples{
# year = year que queremos visualizar
# Get las estadisticas de pitcheo de 1873
 \donttest{get_reference_team_standings(2021)}

}
