% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_team_stats.R
\name{mlb_team_stats}
\alias{mlb_team_stats}
\title{Estadistica oficial bateo y pitcheo de los equipos de la MLB (baseball)}
\usage{
mlb_team_stats(year = year_actual, stats = "hitting", season_type = "regular")
}
\arguments{
\item{year}{numeric}

\item{stats}{character}

\item{season_type}{character}
}
\value{
Estadisticas de la pagina oficial mlb.com
}
\description{
Estadistica oficial bateo y pitcheo de los equipos de la MLB (baseball)
}
\examples{
# year = year que queremos visualizar
# stats = hitting o pitching
# season_type = 'regular', 'playoffs', 'wildcard',
# 'divisionales', 'championship','mundiales' o 'pretemporada
# Get las estadisticas de pitcheo del 2018
# mlb_team_stats(2018, "pitching", "regular")

}
