% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ivo.R
\name{theme_ivo}
\alias{theme_ivo}
\title{Apply ivo theme to a plots}
\usage{
theme_ivo(data, ...)
}
\arguments{
\item{data}{An existing plot object}

\item{...}{Optional additional arguments to plots}
}
\value{
Returns a plot
}
\description{
Apply ivo theme to a plots
}
\examples{
\donttest{
theme_ivo()
}
}
